% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aorsf-package.R
\docType{package}
\name{aorsf-package}
\alias{aorsf}
\alias{aorsf-package}
\title{aorsf: Accelerated Oblique Random Forests}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fit, interpret, and compute predictions with oblique random forests. Includes support for partial dependence, variable importance, passing customized functions for variable importance and identification of linear combinations of features. Methods for the oblique random survival forest are described in Jaeger et al., (2023) \doi{10.1080/10618600.2023.2231048}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/aorsf}
  \item \url{https://docs.ropensci.org/aorsf/}
  \item Report bugs at \url{https://github.com/ropensci/aorsf/issues/}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

Other contributors:
\itemize{
  \item Nicholas Pajewski [contributor]
  \item Sawyer Welden \email{swelden@wakehealth.edu} [contributor]
  \item Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} [reviewer]
  \item Marvin Wright [reviewer]
  \item Lukas Burk [reviewer]
}

}
\keyword{internal}
