% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{as.data.table.aorsf_summary_uni}
\alias{as.data.table.aorsf_summary_uni}
\title{Coerce to data.table}
\usage{
\method{as.data.table}{aorsf_summary_uni}(x, ...)
}
\arguments{
\item{x}{an object of class 'aorsf_summary_uni'}

\item{...}{not used}
}
\value{
a \link[data.table:data.table]{data.table}
}
\description{
Convert an 'aorsf_summary' object into a \code{data.table} object.
}
\examples{

library(data.table)

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id)

smry <- orsf_summarize_uni(object, n_variables = 3)

as.data.table(smry)

}
