% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{as.data.table.orsf_summary_uni}
\alias{as.data.table.orsf_summary_uni}
\title{Coerce to data.table}
\usage{
\method{as.data.table}{orsf_summary_uni}(x, ...)
}
\arguments{
\item{x}{an object of class 'orsf_summary_uni'}

\item{...}{not used}
}
\value{
a \link[data.table:data.table]{data.table}
}
\description{
Convert an 'orsf_summary' object into a \code{data.table} object.
}
\examples{

\dontrun{

library(data.table)

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id, n_tree = 25)

smry <- orsf_summarize_uni(object, n_variables = 2)

as.data.table(smry)

}


}
