% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{One hot encoding}
\usage{
one_hot(data, fi, names_x_data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a \code{data.table} with one-hot encoded factors.
}
\description{
A faster implementation of \link[stats:model.frame]{stats::model.frame} with less options.
}
\note{
One-hot-encoding converts an unordered categorical vector
(i.e. a factor) to multiple binarized vectors where each binary
vector of 1s and 0s indicates the presence of a class (i.e. level)
of the of the original vector.
}
\examples{
n <- 10

data <- data.frame(
  V1 = seq(n),
  V2 = factor(sample(letters[1:3], n, replace = TRUE)),
  V3 = seq(n) / 10,
  V4 = factor(sample(letters[5:6], n, replace = TRUE))
)

data$V1[1] <- NA
data$V3[c(6,7)] <- NA
data$V2[1:2] <- NA
data$V4[2] <- NA

one_hot(data)
}
