% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf2-package.R
\docType{package}
\name{orsf2-package}
\alias{orsf2}
\alias{orsf2-package}
\title{orsf2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/orsf2}
  \item Report bugs at \url{https://github.com/bcjaeger/orsf2/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
