% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_control.R
\name{orsf_control_net}
\alias{orsf_control_net}
\title{ORSF; Penalized regression control}
\usage{
orsf_control_net(alpha = 1/2, df_target = NULL)
}
\arguments{
\item{alpha}{The elastic net mixing parameter. A value of 1 gives the
lasso penalty, and a value of 0 gives the ridge penalty. If multiple
values of alpha are given, then a penalized model is fit using each
alpha value prior to splitting a node.}

\item{df_target}{Preferred number of variables used in a linear combination.
Note: this has to be less than \code{mtry}, which is a separate argument in
\link{orsf} that indicates the number of variables chosen at random prior to
finding a linear combination of those variables.}
}
\value{
an object of class \code{'aorsf_control'}, which should be used as
an input for the \code{control} argument of \link{orsf}.
}
\description{
ORSF; Penalized regression control
}
