% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_control.R
\name{orsf_control_rlt}
\alias{orsf_control_rlt}
\title{ORSF; Reinforcement learning control}
\usage{
orsf_control_rlt(keep = NULL, trees = 35)
}
\arguments{
\item{trees}{}
}
\value{

}
\description{
ORSF; Reinforcement learning control
}
