% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_pd.R
\name{orsf_ice_oob}
\alias{orsf_ice_oob}
\alias{orsf_ice_inb}
\alias{orsf_ice_new}
\title{ORSF Individual Conditional Expectations}
\usage{
orsf_ice_oob(
  object,
  pd_spec,
  pred_horizon = NULL,
  pred_type = "risk",
  expand_grid = TRUE,
  boundary_checks = TRUE,
  ...
)

orsf_ice_inb(
  object,
  pd_spec,
  pred_horizon = NULL,
  pred_type = "risk",
  expand_grid = TRUE,
  boundary_checks = TRUE,
  ...
)

orsf_ice_new(
  object,
  pd_spec,
  new_data,
  pred_horizon = NULL,
  pred_type = "risk",
  expand_grid = TRUE,
  boundary_checks = TRUE,
  ...
)
}
\arguments{
\item{object}{(\emph{orsf_fit}) a trained oblique random survival forest
(see \link{orsf}).}

\item{pd_spec}{(\emph{named list} or \emph{data.frame}).
\itemize{
\item If \code{pd_spec} is a named list,
Each item in the list should be a vector of values that will be used as
points in the partial dependence function. The name of each item in the
list should indicate which variable will be modified to take the
corresponding values.
\item If \code{pd_spec} is a \code{data.frame}, columns will
indicate variable names, values will indicate variable values, and
partial dependence will be computed using the inputs on each row.
}}

\item{pred_horizon}{(\emph{double}) a value or vector indicating the time(s)
that predictions will be calibrated to. E.g., if you were predicting
risk of incident heart failure within the next 10 years, then
\code{pred_horizon = 10}}

\item{pred_type}{(\emph{character}) the type of predictions to compute. Valid
options are
\itemize{
\item 'risk' : probability of having an event at or before \code{pred_horizon}.
\item 'survival' : 1 - risk.
}}

\item{expand_grid}{(\emph{logical}) if \code{TRUE}, partial dependence will be
computed at all possible combinations of inputs in \code{pd_spec}. If
\code{FALSE}, partial dependence will be computed for each variable
in \code{pd_spec}, separately.}

\item{boundary_checks}{(\emph{logical}) if \code{TRUE}, \code{pd_spec} will be vetted
to make sure the requested values are between the 10th and 90th
percentile in the object's training data. If \code{FALSE}, these checks are
skipped.}

\item{...}{Further arguments passed to or from other methods (not currently used).}

\item{new_data}{a \link{data.frame}, \link[tibble:tibble-package]{tibble}, or \link[data.table:data.table]{data.table} to compute predictions in.
Missing data are not currently allowed}
}
\value{
a \link[data.table:data.table]{data.table} containing
individual conditional expectations for the specified variable(s) at the
specified prediction horizon(s).
}
\description{
Compute individual conditional expectations for an ORSF model.
Unlike partial dependence, which shows the expected prediction as a function of one or multiple predictors, individual conditional expectations shows the prediction for an individual observation as a function of a predictor.
}
\details{
You can compute individual conditional expectations two ways using a random forest:
\itemize{
\item using out-of-bag predictions for the training data
\item using predictions for a new set of data
}

See examples for more details
}
\section{Examples}{
Begin by fitting an ORSF ensemble:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(aorsf)

set.seed(329730)

index_train <- sample(nrow(pbc_orsf), 150) 

pbc_orsf_train <- pbc_orsf[index_train, ]
pbc_orsf_test <- pbc_orsf[-index_train, ]

fit <- orsf(data = pbc_orsf_train, 
            formula = Surv(time, status) ~ . - id,
            oobag_pred_horizon = 365.25 * 5)
}\if{html}{\out{</div>}}
\subsection{Individual conditional expectations two ways}{

You can compute individual conditional expectations two ways using a
random forest:
\itemize{
\item using out-of-bag predictions for the training data

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pd_spec <- list(bili = seq(1, 6, length.out = 10))

ice_train <- orsf_ice_oob(fit, pd_spec)

ice_train
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{##       pred_horizon id_variable id_row  bili       pred
##              <num>       <int>  <int> <num>      <num>
##    1:      1826.25           1      1     1 0.08384440
##    2:      1826.25           1      2     1 0.05397791
##    3:      1826.25           1      3     1 0.03162691
##    4:      1826.25           1      4     1 0.14654589
##    5:      1826.25           1      5     1 0.01864370
##   ---                                                 
## 1496:      1826.25          10    146     6 0.82951216
## 1497:      1826.25          10    147     6 0.57515421
## 1498:      1826.25          10    148     6 0.27143096
## 1499:      1826.25          10    149     6 0.74275997
## 1500:      1826.25          10    150     6 0.38469959
}\if{html}{\out{</div>}}
\item using predictions for a new set of data

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ice_test <- orsf_ice_new(fit, 
                         new_data = pbc_orsf_test, 
                         pd_spec = pd_spec)

ice_test
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{##       pred_horizon id_variable id_row  bili       pred
##              <num>       <int>  <int> <num>      <num>
##    1:      1826.25           1      1     1 0.81234703
##    2:      1826.25           1      2     1 0.16038926
##    3:      1826.25           1      3     1 0.37223197
##    4:      1826.25           1      4     1 0.08325345
##    5:      1826.25           1      5     1 0.05819043
##   ---                                                 
## 1256:      1826.25          10    122     6 0.60376866
## 1257:      1826.25          10    123     6 0.22676773
## 1258:      1826.25          10    124     6 0.29596899
## 1259:      1826.25          10    125     6 0.34276120
## 1260:      1826.25          10    126     6 0.36370237
}\if{html}{\out{</div>}}
}
}
}

