% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_interaction.R
\name{orsf_interaction}
\alias{orsf_interaction}
\title{ORSF interactions}
\usage{
orsf_interaction(object, min_pairwise_obs = NULL)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{min_pairwise_obs}{(\emph{integer}) minimum number of observations
where both variables were included in a linear combination together.
The default is the number of trees in \code{object} divided by the mean
number of leaves in the trees.}
}
\value{
a \code{data.frame} with pairwise interaction scores for each
pair of predictor variables in \code{object}.
}
\description{
ORSF interactions
}
\examples{

set.seed(32987)

fit <- orsf(pbc_orsf, Surv(time, status) ~ . - id, n_tree = 2500)

intr <- orsf_interaction(fit)

pd_spec <- list(ascites = c("0","1"),
                bili = seq(0.6, 7.1, by = 0.25))

pd_data <- orsf_pd_summary(fit, pd_spec = pd_spec, times = 1000)

# aligning predictions at lowest value of bili
min_asc_0 <- with(pd_data, mean[ascites == 0 & bili == 0.6])
min_asc_1 <- with(pd_data, mean[ascites == 1 & bili == 0.6])

pd_data_aligned <-
 within(pd_data, {
  mean[ascites == 0] <- mean[ascites == 0] - min_asc_0
  mean[ascites == 1] <- mean[ascites == 1] - min_asc_1
 })

library(ggplot2)

ggplot(pd_data_aligned) +
 aes(x = bili, y = mean, col = factor(ascites)) +
 geom_line() +
 theme_bw()
}
