% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ostree_predict.R
\name{ostree_pred_leaf}
\alias{ostree_pred_leaf}
\title{Title}
\usage{
ostree_pred_leaf(tree, x_new)
}
\arguments{
\item{x_new}{}
}
\value{

}
\description{
Title
}
\examples{

x_new <- flchain_x[1:10, ]

}
