% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ostree_predict.R
\name{ostree_predict}
\alias{ostree_predict}
\title{Title}
\usage{
ostree_predict(object, x_new, x_transforms, times)
}
\arguments{
\item{x_new}{}
}
\value{

}
\description{
Title
}
\examples{

x_new <- flchain_x[1:10, ]

}
