% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{print.aorsf_summary_uni}
\alias{print.aorsf_summary_uni}
\title{Print ORSF summary}
\usage{
\method{print}{aorsf_summary_uni}(x, n_variables = NULL, ...)
}
\arguments{
\item{x}{an object of class 'aorsf_summary'}

\item{n_variables}{The number of variables to print}

\item{...}{not used}
}
\value{
nothing - output is printed to console.
}
\description{
Print ORSF summary
}
\examples{

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id)

summary(object)

}
