% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_print.R
\name{print.orsf_fit}
\alias{print.orsf_fit}
\title{ORSF presentation}
\usage{
\method{print}{orsf_fit}(x, ...)
}
\arguments{
\item{x}{(\emph{orsf_fit}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{...}{Further arguments passed to or from other methods (not currently used).}
}
\value{
nothing - just print output to console
}
\description{
ORSF presentation
}
\examples{

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id, n_tree = 5)

print(object)

}
