% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{print.orsf_summary_uni}
\alias{print.orsf_summary_uni}
\title{Print ORSF summary}
\usage{
\method{print}{orsf_summary_uni}(x, n_variables = NULL, ...)
}
\arguments{
\item{x}{an object of class 'orsf_summary'}

\item{n_variables}{The number of variables to print}

\item{...}{Further arguments passed to or from other methods (not currently used).}
}
\value{
invisibly, \code{x}
}
\description{
Print ORSF summary
}
\examples{

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id, n_tree = 25)

smry <- orsf_summarize_uni(object, n_variables = 2)

print(smry)

}
