% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{summary.aorsf}
\alias{summary.aorsf}
\title{Summarize ORSF model}
\usage{
\method{summary}{aorsf}(object, times, risk = TRUE, ...)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (RSF; see \link{orsf}).}

\item{times}{(\emph{double}) a single time or a vector of times for oblique RSF
predictions. All \code{times} values must not exceed the maximum follow-up
time in the oblique RSF's training data. Also, \code{times} must be entered
in ascending order.}

\item{risk}{(\emph{logical}) if \code{TRUE}, predicted risk is returned. If \code{FALSE},
predicted survival (i.e., 1-risk) is returned.}

\item{...}{not used.}
}
\value{
an object of class 'aorsf_summary'
}
\description{
Summarize ORSF model
}
\examples{

fit <- orsf(pbc_orsf, Surv(time, status) ~ . - id)

summary(fit, times = 1000)

}
