% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aorsf-package.R
\docType{package}
\name{aorsf-package}
\alias{aorsf}
\alias{aorsf-package}
\title{aorsf: Accelerated Oblique Random Survival Forests}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fit, interpret, and make predictions with oblique random survival forests. Oblique decision trees are notoriously slow compared to their axis based counterparts, but 'aorsf' runs as fast or faster than axis-based decision tree algorithms for right-censored time-to-event outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/aorsf}
  \item \url{https://bcjaeger.github.io/aorsf}
  \item Report bugs at \url{https://github.com/bcjaeger/aorsf/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

Other contributors:
\itemize{
  \item Nicholas Pajewski [contributor]
  \item Sawyer Welden \email{swelden@wakehealth.edu} [contributor]
  \item Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} [reviewer]
}

}
\keyword{internal}
