% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_fit.R
\name{orsf}
\alias{orsf}
\title{Title}
\usage{
orsf(
  data,
  formula,
  n_tree = 500,
  n_split = 5,
  mtry = NULL,
  leaf_min_events = 1,
  leaf_min_obs = 15,
  cph_method = "breslow",
  cph_eps = 1e-05,
  cph_iter_max = 1,
  cph_pval_max = 1,
  oobag_pred = FALSE
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms on the right. See details.}

\item{oobag_pred}{}
}
\value{

}
\description{
Title
}
\details{
\emph{formula} The response in \code{formula} can be a survival
object as returned by the \link[survival:Surv]{survival::Surv} function,
but can also just be the time and status variables.
For example, \code{Surv(time, status) ~ .} works just like
\code{time + status ~ .}. The only thing that can break this
input is putting the variables in the wrong order, i.e.,
writing \code{status + time ~ .} will make \code{orsf} assume your
\code{status} variable is actually the \code{time} variable.
}
