% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf2-package.R
\docType{package}
\name{orsf2-package}
\alias{orsf2}
\alias{orsf2-package}
\title{orsf2: Accelerated Oblique Random Survival Forests}
\description{
Oblique random survival forests incorporate linear combinations of input variables into the recursive partitioning algorithm used to grow decision trees. The algorithm used to grow survival trees is the same as in random survival forests from Ishwaran, 2008 <doi:10.1214/08-AOAS169>. A Newton Raphson scoring algorithm based on the Cox proportional hazard model is used to identify linear combinations of input variables. Relative to obliqueRSF, the predecessor of aorsf, decision trees are grown about 500 times faster.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/orsf2}
  \item Report bugs at \url{https://github.com/bcjaeger/orsf2/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
\keyword{internal}
