% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_control.R
\name{orsf_control_cph}
\alias{orsf_control_cph}
\title{ORSF; Newton-Raphson control}
\usage{
orsf_control_cph(
  method = "breslow",
  eps = 1e-05,
  iter_max = 1,
  pval_max = 1,
  do_scale = TRUE
)
}
\arguments{
\item{method}{(\emph{character}) a character string specifying the method
for tie handling. If there are no tied death times all the methods are
equivalent. Valid options are 'breslow' and 'efron'.}

\item{eps}{(\emph{double}) When using Newton Raphson scoring to identify
linear combinations of inputs, iteration continues in the algorithm
until the relative change in  the log partial likelihood is less than
\code{eps}, or the absolute change is less than \code{sqrt(eps)}. Must be positive.}

\item{iter_max}{(\emph{integer}) When using Newton Raphson scoring to identify
linear combinations of inputs, iteration continues until convergence
(see \code{eps} above) or the number of attempted iterations is equal to
\code{iter_max}.}

\item{pval_max}{(\emph{double}) The maximum p-value allowed for a regression
coefficient to remain non-zero. If the p-value for a given coefficient
is above the maximum, the coefficient is set to zero and the variable
no longer plays a role in the linear combination of inputs. Setting
\code{pval_max} to 1 ensures that every predict gets a non-zero
coefficient in the linear combination of inputs.}

\item{do_scale}{(\emph{logical}) if \code{TRUE}, values of predictors will be
scaled prior to running Newton Raphson scoring. Setting to \code{FALSE} will
reduce computation time but will also make the regression extremely
unstable. Therefore, \code{orsf} will only let you set this input to \code{FALSE}
if you also set \code{iter_max} to 1.}
}
\value{
an object of class \code{'aorsf_control'}, which should be used as
an input for the \code{control} argument of \link{orsf}.
}
\description{
ORSF; Newton-Raphson control
}
\examples{

orsf(data_train = pbc_orsf,
     formula = Surv(time, status) ~ . - id,
     control = orsf_control_cph())

}
