% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_control.R
\name{orsf_control_cph}
\alias{orsf_control_cph}
\alias{orsf_control_fast}
\title{Cox proportional hazards control}
\usage{
orsf_control_cph(
  method = "efron",
  eps = 1e-09,
  iter_max = 20,
  do_scale = TRUE,
  ...
)

orsf_control_fast(iter_max = 1, do_scale = FALSE, ...)
}
\arguments{
\item{method}{(\emph{character}) a character string specifying the method
for tie handling. If there are no ties, all the methods are
equivalent. Valid options are 'breslow' and 'efron'. The Efron
approximation is the default because it is more accurate when dealing
with tied event times and has similar computational efficiency compared
to the Breslow method.}

\item{eps}{(\emph{double}) When using Newton Raphson scoring to identify
linear combinations of inputs, iteration continues in the algorithm
until the relative change in  the log partial likelihood is less than
\code{eps}, or the absolute change is less than \code{sqrt(eps)}. Must be positive.
A default value of 1e-09 is used for consistency with
\link[survival:coxph.control]{survival::coxph.control}.}

\item{iter_max}{(\emph{integer}) When using Newton Raphson scoring to identify
linear combinations of inputs, iteration continues until convergence
(see \code{eps} above) or the number of attempted iterations is equal to
\code{iter_max}. A default value of 1 is used for computational efficiency.}

\item{do_scale}{(\emph{logical}) if \code{TRUE}, values of predictors will be
scaled prior to running Newton Raphson scoring. Setting to \code{FALSE} will
reduce computation time but will also make the regression unstable, which
is why the default value is \code{TRUE}. For stability, \code{orsf} will only let
you set this input to \code{FALSE} if you also set \code{iter_max} to 1.}

\item{...}{Further arguments passed to or from other methods
(not currently used).}
}
\value{
an object of class \code{'aorsf_control'}, which should be used as
an input for the \code{control} argument of \link{orsf}.
}
\description{
Use Newton-Raphson scoring to identify linear combinations of input
variables while fitting an \link{orsf} model. For more details on
of Newton-Raphson scoring and the Cox proportional hazards
model, see Therneau and Grambsch (2000).
}
\examples{

orsf(data = pbc_orsf,
     formula = Surv(time, status) ~ . - id,
     control = orsf_control_cph())

}
\references{
Therneau T.M., Grambsch P.M. (2000) The Cox Model. In: Modeling Survival
Data: Extending the Cox Model. Statistics for Biology and Health.
Springer, New York, NY. DOI: 10.1007/978-1-4757-3294-8_3
}
\seealso{
linear combination control functions
\code{\link{orsf_control_custom}()},
\code{\link{orsf_control_net}()}
}
\concept{orsf_control}
