% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{orsf_summarize_uni}
\alias{orsf_summarize_uni}
\title{ORSF summary of univariate information}
\usage{
orsf_summarize_uni(object, n_variables = NULL, times = NULL, risk = TRUE)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{n_variables}{(\emph{integer}) how many variables should be summarized?
Setting this input to a lower number will improve computation time.}

\item{times}{(\emph{double}) a single time or a vector of times for oblique RSF
predictions. All \code{times} values must not exceed the maximum follow-up
time in the oblique RSF's training data. Also, \code{times} must be entered
in ascending order.}

\item{risk}{(\emph{logical}) if \code{TRUE}, predicted risk is returned. If \code{FALSE},
predicted survival (i.e., 1-risk) is returned.}
}
\value{
an object of class 'aorsf_summary'
}
\description{
ORSF summary of univariate information
}
\examples{

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id)

orsf_summarize_uni(object, n_variables = 3)

}
