% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{orsf_summarize_uni}
\alias{orsf_summarize_uni}
\title{ORSF summary of univariate information}
\usage{
orsf_summarize_uni(
  object,
  n_variables = NULL,
  pred_horizon = NULL,
  pred_type = "risk",
  importance = "negate",
  ...
)
}
\arguments{
\item{object}{an untrained aorsf object, created by setting
\code{no_fit = TRUE} in \code{orsf()}.}

\item{n_variables}{(\emph{integer}) how many variables should be summarized?
Setting this input to a lower number will reduce computation time.}

\item{pred_horizon}{(\emph{double}) a single time or a vector of times
indicating the time(s) that predicted risk or survival probabilities
will be computed at.}

\item{pred_type}{(\emph{character}) the type of predictions to return. Valid
options are
\itemize{
\item 'risk' : probability of having an event at or before \code{pred_horizon}.
\item 'survival' : 1 - risk.
}}

\item{importance}{(\emph{character}) Indicate method for variable importance:
\itemize{
\item 'none': no variable importance is computed.
\item 'anova': use the analysis of variance (ANOVA) method
\item 'negate': compute negation importance
\item 'permute': compute permutation importance
}}

\item{...}{not currently used.}
}
\value{
an object of class 'aorsf_summary'
}
\description{
ORSF summary of univariate information
}
\details{
If \code{pred_horizon} is left unspecified, the median value of
the time-to-event variable in \code{object}'s training data will be used.
It is recommended to always specify your own prediction horizon,
as the median time may not be an especially meaningful horizon to
compute predicted risk values at.

If \code{object} already has variable importance values, you can
safely bypass the computation of variable importance in this function
by setting importance = 'none'.
}
\examples{

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id)

# since anova importance was used to make object, we can
# safely say importance = 'none' and skip computation of
# variable importance while running orsf_summarize_uni

orsf_summarize_uni(object, n_variables = 3, importance = 'none')

# however, if we want to summarize object according to variables
# ranked by negation importance, we can compute negation importance
# within orsf_summarize_uni() as follows:

orsf_summarize_uni(object, n_variables = 3, importance = 'negate')


}
