% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_summary.R
\name{orsf_summarize_uni}
\alias{orsf_summarize_uni}
\title{ORSF summary of univariate information}
\usage{
orsf_summarize_uni(
  object,
  n_variables = NULL,
  pred_horizon = NULL,
  risk = TRUE,
  importance_type = "negate"
)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{n_variables}{(\emph{integer}) how many variables should be summarized?
Setting this input to a lower number will improve computation time.}

\item{pred_horizon}{(\emph{double}) a single time or a vector of times
indicating the prediction horizon. Predicted risk or survival values
will indicate the probability of having an event or surviving from
baseline to the prediction horizon, respectively. All \code{pred_horizon}
values must not exceed the maximum follow-up time in \code{object}'s
training data. Also, \code{pred_horizon} values must be entered in
ascending order.}

\item{risk}{(\emph{logical}) if \code{TRUE}, predicted risk is returned. If \code{FALSE},
predicted survival (i.e., 1-risk) is returned.}

\item{importance_type}{(\emph{character}) which method to use for variable
importance. Valid input values are 'anova' and 'negate'.
See \link{orsf_vi_negate} for details on these methods.}
}
\value{
an object of class 'aorsf_summary'
}
\description{
ORSF summary of univariate information
}
\details{
if \code{pred_horizon} is left unspecified, the median value of
the time-to-event variable in \code{object}'s training data will be used.
It is recommended to always specify your own prediction horizon,
as the median time may not be an especially meaningful horizon to
compute predicted risk values at.
}
\examples{

object <- orsf(pbc_orsf, Surv(time, status) ~ . - id)

orsf_summarize_uni(object, n_variables = 3)

}
