% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_vi.R
\name{orsf_vi}
\alias{orsf_vi}
\title{ORSF variable importance}
\usage{
orsf_vi(object)
}
\arguments{
\item{object}{an object of class 'aorsf'.}
}
\value{
the \code{object} with variable importance attached.
}
\description{
ORSF variable importance
}
\examples{

fit <- orsf(pbc_orsf,
            Surv(time, status) ~ . - id,
            oobag_pred = TRUE)

orsf_vi(fit)

}
