% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_vi.R
\name{orsf_vi}
\alias{orsf_vi}
\title{ORSF variable importance}
\usage{
orsf_vi(object, group_factors = TRUE)
}
\arguments{
\item{object}{an object of class 'aorsf'.}

\item{group_factors}{(\emph{logical}) if \code{TRUE}, the importance of factor
variables will be reported overall by aggregating the importance
of individual levels of the factor. If \code{FALSE}, the importance of
individual factor levels will be returned.}
}
\value{
the \code{object} with variable importance attached.
}
\description{
ORSF variable importance
}
\details{
The method used to compute variable importance with ORSF is called
'negation importance'. Each variable is assessed separately by
multiplying the variable's coefficients by -1 and then determining
how much the model's performance changes. The worse the model's
performance after negating all coefficients for a given variable,
the more important the variable.
}
\examples{

fit <- orsf(pbc_orsf,
            Surv(time, status) ~ . - id,
            oobag_pred = TRUE)

orsf_vi(fit)

}
