% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_vi.R
\name{orsf_vi_negate}
\alias{orsf_vi_negate}
\alias{orsf_vi_anova}
\title{ORSF variable importance}
\usage{
orsf_vi_negate(object, group_factors = TRUE)

orsf_vi_anova(object, group_factors = TRUE)
}
\arguments{
\item{object}{an object of class 'aorsf'.}

\item{group_factors}{(\emph{logical}) if \code{TRUE}, the importance of factor
variables will be reported overall by aggregating the importance
of individual levels of the factor. If \code{FALSE}, the importance of
individual factor levels will be returned.}
}
\value{
a named vector. Names indicate predictors, values indicate importance.
The vector is sorted from highest to lowest value, with higher values
indicating higher importance.
}
\description{
Determine the importance of individual variables using 'negation
importance.' See 'Details' for definition of negation importance.
}
\details{
\strong{negation importance}: Each variable is assessed separately by
multiplying the variable's coefficients by -1 and then determining
how much the model's performance changes. The worse the model's
performance after negating coefficients for a given variable,
the more important the variable.

\strong{ANOVA importance}: ANOVA importance computes a p-value for each
coefficient in each linear combination of variables in each decision
tree of an oRF. Following the definition proposed by Menze et al,
ANOVA importance in aorsf for an individual variable is the proportion
of times a p-value for its coefficient is < 0.10.
}
\examples{

fit <- orsf(pbc_orsf,
            Surv(time, status) ~ . - id,
            oobag_pred = TRUE)

orsf_vi_negate(fit)
orsf_vi_anova(fit)

}
\references{
Menze, Bjoern H., et al. On oblique random forests.
\emph{Joint European Conference on Machine Learning and Knowledge Discovery in Databases}.
Springer, Berlin, Heidelberg, 2011.
DOI: 10.1007/978-3-642-23783-6_29
}
