% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_vs.R
\name{orsf_vs}
\alias{orsf_vs}
\title{Variable selection}
\usage{
orsf_vs(object, n_predictor_min = 3, verbose_progress = FALSE)
}
\arguments{
\item{object}{(\emph{orsf_fit}) a trained oblique random survival forest
(see \link{orsf}).}

\item{n_predictor_min}{(\emph{integer}) the minimum number of predictors allowed}

\item{verbose_progress}{(\emph{logical}) not implemented yet. Should progress be printed to the console?}
}
\value{
a \link[data.table:data.table]{data.table} with four columns:
\itemize{
\item \emph{n_predictors}: the number of predictors used
\item \emph{stat_value}: the out-of-bag statistic
\item \emph{predictors_included}: the names of the predictors included
\item \emph{predictor_dropped}: the predictor selected to be dropped
}
}
\description{
Variable selection
}
\examples{

object <- orsf(formula = time + status ~ .,
               data = pbc_orsf,
               n_tree = 25,
               importance = 'anova',
               tree_seeds = 1:25)

orsf_vs(object)
}
