% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_predict.R
\name{predict.aorsf}
\alias{predict.aorsf}
\title{Predict risk or survival}
\usage{
\method{predict}{aorsf}(object, new_data, pred_horizon, risk = TRUE, ...)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{new_data}{(\emph{data.frame}) data to compute predictions for. Must have
the same columns with equivalent types as the data used to train \code{object}.
Also, factors in \code{new_data} must not have levels that were not in the
data used to train \code{object}. Last, missing data are not supported.}

\item{pred_horizon}{(\emph{double}) a single time indicating the prediction horizon.
Predicted risk or survival values will indicate the probability of
having an event or surviving from baseline to the prediction horizon.
When using \code{\link[=predict.aorsf]{predict.aorsf()}}, \code{pred_horizon} can be a vector of arbitrary length.
When using \code{\link[=orsf_pd_summary]{orsf_pd_summary()}} or \code{\link[=orsf_pd_ice]{orsf_pd_ice()}}, \code{pred_horizon} must be
length 1. All \code{pred_horizon} values must not exceed the maximum follow-up
time in the oblique RSF's training data. Also, \code{pred_horizon} must be entered
in ascending order.}

\item{risk}{(\emph{logical}) if \code{TRUE}, predicted risk is returned. If \code{FALSE},
predicted survival (i.e., 1-risk) is returned.}

\item{...}{not used.}
}
\value{
a \code{matrix} of predictions. Column \code{j} of the matrix corresponds
to value \code{j} in \code{pred_horizon}. Row \code{i} of the matrix corresponds to row \code{i}
in \code{new_data}.
}
\description{
Predict risk or survival
}
\examples{

train <- seq(1, nrow(pbc_orsf), by = 2)
test <- seq(2, nrow(pbc_orsf), by = 2)

fit <- orsf(pbc_orsf[train, ], Surv(time, status) ~ . - id)

preds <- predict(fit,
                 new_data = pbc_orsf[test, ],
                 pred_horizon = c(500, 1500, 2500))

head(preds)


}
