% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_predict.R
\name{predict.aorsf}
\alias{predict.aorsf}
\title{ORSF predictions}
\usage{
\method{predict}{aorsf}(object, new_data, times, risk = TRUE, ...)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{new_data}{(\emph{data.frame}) data to compute predictions for. Must have
the same columns with equivalent types as the data used to train \code{object}.
Also, factors in \code{new_data} must not have levels that were not in the
data used to train \code{object}. Last, missing data are not supported.}

\item{times}{(\emph{double}) a single time or a vector of times for oblique RSF
predictions. All \code{times} values must not exceed the maximum follow-up
time in the oblique RSF's training data. Also, \code{times} must be entered
in ascending order.}

\item{risk}{(\emph{logical}) if \code{TRUE}, predicted risk is returned. If \code{FALSE},
predicted survival (i.e., 1-risk) is returned.}

\item{...}{not used.}
}
\value{
a \code{matrix} of predictions. Column \code{j} of the matrix corresponds
to value \code{j} in \code{times}. Row \code{i} of the matrix corresponds to row \code{i}
in \code{new_data}.
}
\description{
ORSF predictions
}
\examples{

train <- seq(1, nrow(pbc_orsf), by = 2)
test <- seq(2, nrow(pbc_orsf), by = 2)

fit <- orsf(pbc_orsf[train, ], Surv(time, status) ~ . - id)

preds <- predict(fit,
                 new_data = pbc_orsf[test, ],
                 times = c(500, 1500, 2500))

head(preds)


}
