% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_predict.R
\name{predict.aorsf}
\alias{predict.aorsf}
\title{Predict using oblique random survival forest}
\usage{
\method{predict}{aorsf}(object, new_data, pred_horizon, pred_type = "risk", ...)
}
\arguments{
\item{object}{(\emph{aorsf}) an oblique random survival forest (ORSF; see \link{orsf}).}

\item{new_data}{(\emph{data.frame}) data to compute predictions for. Must have
the same columns with equivalent types as the data used to train \code{object}.
Also, factors in \code{new_data} must not have levels that were not in the
data used to train \code{object}. Last, missing data are not supported.}

\item{pred_horizon}{(\emph{double}) a single time or a vector of times
indicating the time(s) that predicted risk or survival probabilities
will be computed at.}

\item{pred_type}{(\emph{character}) the type of predictions to return. Valid
options are
\itemize{
\item 'risk' : probability of having an event at or before \code{pred_horizon}.
\item 'survival' : 1 - risk.
}}

\item{...}{Further arguments passed to or from other methods
(not currently used).}
}
\value{
a \code{matrix} of predictions. Column \code{j} of the matrix corresponds
to value \code{j} in \code{pred_horizon}. Row \code{i} of the matrix corresponds to
row \code{i} in \code{new_data}.
}
\description{
Predict using oblique random survival forest
}
\details{
\code{pred_horizon} values must not exceed the maximum follow-up time in
\code{object}'s training data. Also, \code{pred_horizon} values must be entered
in ascending order.
}
\examples{


# indices of data used for training the model
train <- seq(1, nrow(pbc_orsf), by = 2)

# indices of data used to test the trained model.
test <- seq(2, nrow(pbc_orsf), by = 2)

fit <- orsf(pbc_orsf[train, ], Surv(time, status) ~ . - id)

preds <- predict(fit,
                 new_data = pbc_orsf[test, ],
                 pred_horizon = c(500, 1500, 2500))

head(preds)


}
