% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsf_print.R
\name{print.aorsf}
\alias{print.aorsf}
\title{ORSF presentation}
\usage{
\method{print}{aorsf}(x, ...)
}
\arguments{
\item{x}{an object of class 'aorsf'}

\item{...}{not used}
}
\value{
nothing - just print output to console
}
\description{
ORSF presentation
}
